README.md
This is my first git project. It is not a dummy project. I'm trying to make a program that renames multiple files with the right number format. It was initially designed to rename multiple srt files so My video player would automatically put the subtitle on the video. I am using this opportunity to learn how to create a fully functioning program that has UI and everything. But thats a long way ahead. one day for sure! -Mazdak